auth_service/requirements.txt
django==5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
mysqlclient==2.2.7
bcrypt==5.0.0
pytz
python-decouple
gunicorn